+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - User Endpoint
+- Sidebar option to invert columns order. 
 
 ### Changed
 - Refactor /admin/users views using a Presenter.
@@ -27,4 +28,3 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 [Unreleased]: https://github.com/Codeminer42/cm42-central/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/Codeminer42/cm42-central/tree/v1.0.0
-